langchain-edenmarco-udemy-ws/langchain-agentic/pyproject.toml
[project]
name = "langchain-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-groq>=0.1.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
]
